requirements.txt
faiss-cpu==1.9.0.post1
fastapi==0.115.6
numpy==1.26.0
sentence-transformers==3.3.1
uvicorn==0.34.0
pydantic>=2.1.1,<3.0
